refactor(ProductController.java): separate service logic
